# gateway-showcase/api/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Admin API
  description: Enterprise API Gateway Management Platform API Documentation
  version: 1.0.0
  contact:
    name: Your Name
    email: your-email@example.com

servers:
  - url: http://localhost:9090
    description: Development server

tags:
  - name: Routes
    description: Route management operations
  - name: Services
    description: Service management operations
  - name: Strategies
    description: Strategy management operations
  - name: Monitoring
    description: Monitoring and metrics operations
  - name: Alerts
    description: Alert configuration and history

paths:
  /api/routes:
    get:
      tags:
        - Routes
      summary: List all routes
      description: Retrieve a paginated list of all gateway routes
      operationId: listRoutes
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
    post:
      tags:
        - Routes
      summary: Create a route
      description: Create a new gateway route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreateRequest'
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/routes/{routeId}:
    get:
      tags:
        - Routes
      summary: Get route by ID
      operationId: getRoute
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '404':
          description: Route not found
    put:
      tags:
        - Routes
      summary: Update a route
      operationId: updateRoute
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteUpdateRequest'
      responses:
        '200':
          description: Route updated successfully
    delete:
      tags:
        - Routes
      summary: Delete a route
      operationId: deleteRoute
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Route deleted successfully

  /api/services:
    get:
      tags:
        - Services
      summary: List all services
      operationId: listServices
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'

  /api/strategies:
    get:
      tags:
        - Strategies
      summary: List all strategies
      operationId: listStrategies
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'

  /api/monitor/metrics:
    get:
      tags:
        - Monitoring
      summary: Get real-time metrics
      operationId: getMetrics
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'

  /api/alerts/history:
    get:
      tags:
        - Alerts
      summary: Get alert history
      operationId: getAlertHistory
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'

components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object

    PageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            content:
              type: array
              items:
                type: object
            totalElements:
              type: integer
            totalPages:
              type: integer
            page:
              type: integer
            size:
              type: integer

    RouteCreateRequest:
      type: object
      required:
        - routeName
        - uri
      properties:
        routeName:
          type: string
          example: User Service Route
        uri:
          type: string
          example: lb://user-service
        predicates:
          type: array
          items:
            $ref: '#/components/schemas/PredicateDefinition'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/FilterDefinition'
        order:
          type: integer
          default: 0
        enabled:
          type: boolean
          default: true

    RouteUpdateRequest:
      type: object
      properties:
        routeName:
          type: string
        uri:
          type: string
        predicates:
          type: array
          items:
            $ref: '#/components/schemas/PredicateDefinition'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/FilterDefinition'
        order:
          type: integer
        enabled:
          type: boolean

    RouteResponse:
      type: object
      properties:
        routeId:
          type: string
        routeName:
          type: string
        uri:
          type: string
        predicates:
          type: array
          items:
            $ref: '#/components/schemas/PredicateDefinition'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/FilterDefinition'
        order:
          type: integer
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PredicateDefinition:
      type: object
      properties:
        name:
          type: string
          example: Path
        args:
          type: object
          additionalProperties:
            type: string
          example:
            pattern: /api/user/**

    FilterDefinition:
      type: object
      properties:
        name:
          type: string
          example: StripPrefix
        args:
          type: object
          additionalProperties:
            type: string
          example:
            parts: "1"

    MetricsResponse:
      type: object
      properties:
        code:
          type: integer
        data:
          type: object
          properties:
            processCpuUsage:
              type: number
              format: double
            systemCpuUsage:
              type: number
              format: double
            heapUsagePercent:
              type: number
              format: double
            totalRequests:
              type: integer
            errorRate:
              type: number
              format: double
            avgResponseTimeMs:
              type: number
              format: double

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []